Extract MSIDXpcCanPerformFailureReason into a lightweight header
The two Xpc token request controller public headers imported the full
MSIDXpcSingleSignOnProvider.h solely to reference the
MSIDXpcCanPerformFailureReason enum. That transitively pulled the provider
interface and its MSIDSSOExtensionRequestDelegate dependency into every
downstream consumer of the controller headers (e.g. OneAuth's
MSAIBrokerClient.m), breaking the "Common core subtree update check macOS"
build with "expected a type" on MSIDSSOExtensionRequestDelegateCompletionBlock.

Move the enum and its ToString declaration into a new standalone
MSIDXpcCanPerformFailureReason.h (Foundation-only). The provider header and
both controller headers now import this lightweight header; the controller .m
files continue to import the full provider header for the class method call.
No public API or behavior change.

Copilot-Session: fcb8c919-d264-449a-ba3e-dfdd541c2b9f
