+++ IdentityCore/src/controllers/broker/mac/MSIDXpcSilentTokenRequestController.h
@@ -24,7 +24,7 @@
 
 
 #import "MSIDSilentController.h"
-#import "MSIDXpcSingleSignOnProvider.h"
+#import "MSIDXpcCanPerformFailureReason.h"
 
 NS_ASSUME_NONNULL_BEGIN
