+++ IdentityCore/src/util/mac/MSIDXpcSingleSignOnProvider.h
@@ -27,31 +27,9 @@
 #import "MSIDSSOExtensionRequestDelegate.h"
 #import "MSIDRequestContext.h"
 #import "MSIDXpcProviderCaching.h"
+#import "MSIDXpcCanPerformFailureReason.h"
 
 NS_ASSUME_NONNULL_BEGIN
-
-// Distinguishes why +[MSIDXpcSingleSignOnProvider canPerformRequest:] returned NO,
-// so telemetry/logging can disambiguate which of the client-side static gates rejected the request.
-typedef NS_ENUM(NSInteger, MSIDXpcCanPerformFailureReason)
-{
-    // canPerformRequest succeeded, no failure occurred.
-    MSIDXpcCanPerformFailureReasonNone = 0,
-    // Neither the MacBrokerApp nor the CompanyPortal Xpc component is installed on the device.
-    MSIDXpcCanPerformFailureReasonNoProviderInstalled,
-    // Failed to construct the SSOExtension getDeviceInfo request object.
-    MSIDXpcCanPerformFailureReasonDeviceInfoRequestCreationFailed,
-    // SSOExtension getDeviceInfo handshake completed with a hard error.
-    MSIDXpcCanPerformFailureReasonDeviceInfoHandshakeError,
-    // SSOExtension getDeviceInfo handshake did not complete before the 1 second timeout expired.
-    MSIDXpcCanPerformFailureReasonDeviceInfoHandshakeTimeout,
-    // No installed Xpc provider matches the cached/available Xpc configuration.
-    MSIDXpcCanPerformFailureReasonValidateCacheProviderFailed,
-    // The Xpc broker flow is gated behind macOS 13+ (or otherwise unsupported on this OS version) and was rejected before reaching MSIDXpcSingleSignOnProvider.
-    MSIDXpcCanPerformFailureReasonUnsupportedOSVersion,
-};
-
-// Returns a human readable, non-PII name for the given failure reason, suitable for logging.
-extern NSString *MSIDXpcCanPerformFailureReasonToString(MSIDXpcCanPerformFailureReason reason);
 
 @interface MSIDXpcSingleSignOnProvider : NSObject
