+++ changelog.txt
@@ -1,5 +1,5 @@
 TBD
-* Add MSIDXpcCanPerformFailureReason to +[MSIDXpcSingleSignOnProvider canPerformRequest:reason:], disambiguating why the client-side Xpc gate rejected a request (no provider installed, getDeviceInfo request creation failure, handshake error, handshake timeout, or cache validation failure), and log the reason from the silent/interactive Xpc controllers. Add matching +[MSIDXpcSilentTokenRequestController canPerformRequest:] and +[MSIDXpcInteractiveTokenRequestController canPerformRequest:] overloads (with a new UnsupportedOSVersion reason for the pre-macOS-13 gate) so external callers can surface the failure reason to their own telemetry.
+* Add MSIDXpcCanPerformFailureReason to +[MSIDXpcSingleSignOnProvider canPerformRequest:reason:], disambiguating why the client-side Xpc gate rejected a request (no provider installed, getDeviceInfo request creation failure, handshake error, handshake timeout, or cache validation failure), and log the reason from the silent/interactive Xpc controllers. Add matching +[MSIDXpcSilentTokenRequestController canPerformRequest:] and +[MSIDXpcInteractiveTokenRequestController canPerformRequest:] overloads (with a new UnsupportedOSVersion reason for the pre-macOS-13 gate) so external callers can surface the failure reason to their own telemetry. The reason enum lives in a lightweight MSIDXpcCanPerformFailureReason.h so the controller headers no longer pull the full MSIDXpcSingleSignOnProvider interface (and its SSOExtension dependencies) into downstream consumers.
 * Complete customized macOS XPC transport failures exactly once with non-nil errors, including bounded dispatcher endpoint and broker reply timeouts, missing connection/proxy/endpoint handling, and synchronous interactive dispatch failures.
 * Gate the reqCnf presence validation for Pop token requests in MSIDBrowserNativeMessageGetTokenRequest behind the browser_core_disable_reqcnf_validation flight (kill switch, validation enabled by default).
 * iOS: opt-in test injection hook for in-process CBA challenge handling #1889
